refactor: move pure helper modules from services/ to utils/
