+++ src/main/src/utils/coreScript.ts
