+++ src/main/src/utils/platformTransfer.ts
