+++ src/main/src/utils/shieldedNoteSelection.ts
