+++ tests/unit/platformInputSelection.test.ts
@@ -4,7 +4,7 @@ import {
   PlatformSourceCandidate,
   MIN_INPUT_CREDITS,
   MAX_ADDRESS_INPUTS,
-} from '../../src/main/src/services/platformTransfer'
+} from '../../src/main/src/utils/platformTransfer'
 
 function candidate(platformAddress: string, balanceCredits: bigint, nonce = 0): PlatformSourceCandidate {
   return {
